packages/shared-py/pyproject.toml
[project]
name = "annemusic-shared"
version = "0.1.0"
description = "Shared Python library for annemusic stages (logger, redis, gcs, env)"
requires-python = ">=3.11"
dependencies = [
  "upstash-redis>=1.2.0",
  "google-cloud-storage>=2.19.0",
  "google-auth>=2.37.0",
  "jsonschema>=4.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]

[dependency-groups]
dev = ["ruff>=0.7"]
